Filter builder refactor WIP
